improve max_event handling

+++ cluster.cpp
@@ -75,7 +75,7 @@ Cluster& Cluster::operator=(const Cluster& other){
 	angle = other.angle;
 	return *this;
 }
-Cluster::Cluster(T * treeObject, int entry){
+Cluster::Cluster(T * treeObject, long entry){
 	if(entry>-1){
 		treeObject->LoadTree(entry);
 		treeObject->GetEntry(entry);
@@ -178,7 +178,7 @@ CM_Cluster& CM_Cluster::operator=(const CM_Cluster& other){
 	cm_n_in_tree = other.cm_n_in_tree;
 	return *this;
 }
-CM_Cluster::CM_Cluster(T * treeObject,int number_,CM_Detector * det, int entry): Cluster(treeObject,entry){
+CM_Cluster::CM_Cluster(T * treeObject,int number_,CM_Detector * det, long entry): Cluster(treeObject,entry){
 	if(entry>-1){
 		treeObject->LoadTree(entry);
 		treeObject->GetEntry(entry);
@@ -217,7 +217,7 @@ CM_Cluster::CM_Cluster(CM_Detector * det, int number_, double pos_, double size_
 CM_Cluster::~CM_Cluster(){
 	
 }
-bool CM_Cluster::is_suitable(T * treeObject,int number_,CM_Detector * detector, int entry){
+bool CM_Cluster::is_suitable(T * treeObject,int number_,CM_Detector * detector, long entry){
 	//cout << "blah" << endl;
 	if(entry>-1){
 		treeObject->LoadTree(entry);
@@ -398,7 +398,7 @@ MG_Cluster& MG_Cluster::operator=(const MG_Cluster& other){
 	mg_n_in_tree = other.mg_n_in_tree;
 	return *this;
 }
-MG_Cluster::MG_Cluster(T * treeObject,int number_,MG_Detector * det, int entry): Cluster(treeObject,entry){
+MG_Cluster::MG_Cluster(T * treeObject,int number_,MG_Detector * det, long entry): Cluster(treeObject,entry){
 	if(entry>-1){
 		treeObject->LoadTree(entry);
 		treeObject->GetEntry(entry);
@@ -436,7 +436,7 @@ MG_Cluster::MG_Cluster(MG_Detector * det, int number_, double pos_, double size_
 MG_Cluster::~MG_Cluster(){
 	
 }
-bool MG_Cluster::is_suitable(T * treeObject,int number_,MG_Detector * detector, int entry){
+bool MG_Cluster::is_suitable(T * treeObject,int number_,MG_Detector * detector, long entry){
 	if(entry>-1){
 		treeObject->LoadTree(entry);
 		treeObject->GetEntry(entry);
