+++ event.cpp
@@ -101,7 +101,7 @@ Event& Event::operator=(const Event& other){
 	is_X = other.is_X;
 	return *this;
 }
-Event::Event(T * treeObject, bool use_srf_,int entry){
+Event::Event(T * treeObject, bool use_srf_,long entry){
 	if(entry>-1){
 		treeObject->LoadTree(entry);
 		treeObject->GetEntry(entry);
@@ -157,7 +157,7 @@ CM_Event& CM_Event::operator=(const CM_Event& other){
 	detector = other.detector;
 	return *this;
 }
-CM_Event::CM_Event(T * treeObject,CM_Detector * det, bool use_srf_,int entry): Event(treeObject,use_srf_,entry){
+CM_Event::CM_Event(T * treeObject,CM_Detector * det, bool use_srf_,long entry): Event(treeObject,use_srf_,entry){
 	if(entry>-1){
 		treeObject->LoadTree(entry);
 		treeObject->GetEntry(entry);
@@ -298,7 +298,7 @@ MG_Event& MG_Event::operator=(const MG_Event& other){
 	detector = other.detector;
 	return *this;
 }
-MG_Event::MG_Event(T * treeObject,MG_Detector * det, bool use_srf_,int entry): Event(treeObject,use_srf_,entry){
+MG_Event::MG_Event(T * treeObject,MG_Detector * det, bool use_srf_,long entry): Event(treeObject,use_srf_,entry){
 	if(entry>-1){
 		treeObject->LoadTree(entry);
 		treeObject->GetEntry(entry);
@@ -730,7 +730,7 @@ CosmicBenchEvent& CosmicBenchEvent::operator=(const CosmicBenchEvent& other){
 	}
 	return *this;
 }
-CosmicBenchEvent::CosmicBenchEvent(CosmicBench * detectors, T * treeObject, bool use_srf_, int entry){
+CosmicBenchEvent::CosmicBenchEvent(CosmicBench * detectors, T * treeObject, bool use_srf_, long entry){
 	if(entry>-1){
 		treeObject->LoadTree(entry);
 		treeObject->GetEntry(entry);
